refactor: use web events everywhere
