+++ packages/fiber/src/index.tsx
@@ -13,9 +13,8 @@ export type {
   RootState,
 } from './core/store'
 export type { ThreeEvent, Events, EventManager, ComputeFunction } from './core/events'
-export { createEvents } from './core/events'
+export { createEvents, createPointerEvents as events } from './core/events'
 export type { ObjectMap, Camera } from './core/utils'
 export * from './web/Canvas'
-export { createPointerEvents as events } from './web/events'
 export type { GlobalRenderCallback, GlobalEffectType } from './core/loop'
 export * from './core'
